+++ CHANGELOG.md
@@ -19,6 +19,10 @@
 * https://github.com/malachi-iot/estdlib/issues/169 fixed glitchy from_chars sto mode hex/octal detection
 * https://github.com/malachi-iot/estdlib/issues/173 additional units operators for 1:1 mode
 
+## Known Issues and Limitations
+
+* https://github.com/malachi-iot/estdlib/issues/177 chrono::duration (for some time now) has too permissive of a constructor
+
 # v0.8.10 - 08OCT25
 
 ## Added Features